Moved injection to ResolveReferences
