--- Defs/ModHelperDefs/ModHelper.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<Defs>
-
-    <CommunityCoreLibrary.ModHelperDef>
-        <defName>Miniaturisation</defName>
-
-        <ModName>Miniaturisation</ModName>
-		
-        <minCCLVersion>0.14.0</minCCLVersion>
-
-        <SpecialInjectors>
-            <li>Miniaturisation.DefInjector</li>
-        </SpecialInjectors>
-
-    </CommunityCoreLibrary.ModHelperDef>
-
-</Defs>
