+++ Source/Miniaturisation.userprefs
@@ -2,8 +2,8 @@
   <MonoDevelop.Ide.Workspace ActiveConfiguration="Release" />
   <MonoDevelop.Ide.Workbench ActiveDocument="DefInjector.cs">
     <Files>
-      <File FileName="DefInjector.cs" Line="53" Column="41" />
-      <File FileName="MiniaturisationDef.cs" Line="14" Column="3" />
+      <File FileName="DefInjector.cs" Line="54" Column="10" />
+      <File FileName="MiniaturisationDef.cs" Line="1" Column="1" />
     </Files>
   </MonoDevelop.Ide.Workbench>
   <MonoDevelop.Ide.DebuggingService.Breakpoints>
